Project-Exam/requirements.txt
Flask==3.0.3
PyMySQL==1.1.1
SQLAlchemy==2.0.36
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.97.0
google-cloud-secret-manager==2.16.3